+++ docs/source/changelog.rst
@@ -5,6 +5,13 @@ Changelog
 	:local:
 	:backlinks: none
 
+-------------
+Version (dev)
+-------------
+* Renamed :py:attr:`!NO_FLASHLIGHT` in :py:attr:`bsp.StaticPropFlags <srctools.bsp.StaticPropFlags>` to 
+  :py:attr:`NO_SHADOW_DEPTH <srctools.bsp.StaticPropFlags.NO_SHADOW_DEPTH>` to reflect the actual 
+  behaviour of the flag, added the real :py:attr:`NO_FLASHLIGHT <srctools.bsp.StaticPropFlags.NO_FLASHLIGHT>` define.
+
 --------------
 Version 2.3.12
 --------------